% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.wgnam.R
\name{print.summary.wgnam}
\alias{print.summary.wgnam}
\title{Print table of QTL information}
\usage{
\method{print}{summary.wgnam}(x, ...)
}
\arguments{
\item{x}{an object of class 'summary.wgnam'}

\item{\ldots}{other arguments passed to summary.wgnam
(ignored)}
}
\description{
Print method for objects of class 'summary.wgnam'
}
\details{
Prints the summary information found using 'summary.wgnam'
}
\author{
Ari Verbyla
}
