% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.wgnam.R
\name{print.wgnam}
\alias{print.wgnam}
\title{print method for 'wgnam'}
\usage{
\method{print}{wgnam}(object, namObj, ...)
}
\arguments{
\item{object}{an object of class 'wgnam'}

\item{namObj}{nam genetic data}

\item{\ldots}{Additional optional objects}
}
\description{
Provides a simple printing of the selected QTL after using 'wgnam'.
}
\author{
Valeria Paccapelo
}
