% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.wgnam.R
\name{summary.wgnam}
\alias{summary.wgnam}
\title{Summary of the QTL analysis after using wgnam}
\usage{
\method{summary}{wgnam}(object, namObj, merge.by = "id", ...)
}
\arguments{
\item{object}{an object of class \code{wgnam}}

\item{namObj}{a list object containing the genotypic data,}

\item{merge.by}{The genetic line variable used in merging
phenotypic and genetic data in the \code{wgnam} analysis.}

\item{\ldots}{Optional additional objects passed to or from other
methods}
}
\value{
A object of class \code{summary.wgnam}, consisting of a list
with a data frame of QTL information.
}
\description{
Performs a summary after a QTL analysis using \code{wgnam}. QTL effects are organised in a presentable format.
}
\examples{
\dontrun{
SumNAM <- summary(wgnam_qtl)
}

}
\author{
Ari Verbyla
}
