% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgnam-package.R
\docType{package}
\name{wgnam-package}
\alias{wgnam-package}
\title{wgnam: Whole Genome Nested Association Mapping for QTL detection using mixed models}
\description{
A computationally efficient whole genome approach to detecting and estimating significant QTL
in Nested Association Mapping Populations using the flexible one-stage linear mixed modelling functionality of ASReml-R.
}
\author{
Ari Verbyla and Valeria Paccapelo
}
