% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgnam.asreml.R
\name{wgnam.asreml}
\alias{wgnam.asreml}
\title{wgnam: Whole Genome Nested Association Mapping for QTL detection and estimation using ASReml-R}
\usage{
\method{wgnam}{asreml}(
  baseModel,
  phenoData,
  namObj,
  merge.by = NULL,
  aped = NULL,
  prop = 1,
  res = FALSE,
  eq.psi = FALSE,
  TypeI = 0.05,
  attempts = 5,
  breakout = -1,
  data.name = NULL,
  trace = TRUE,
  verboseLev = 0,
  ...
)
}
\arguments{
\item{baseModel}{a model object of class \code{asreml} usually
representing a base model with which to build the model.}

\item{phenoData}{a data frame containing the phenotypic elements
used to fit \code{baseModel}. This data is checked against the data
used in fitting the base model}

\item{namObj}{a list object containing the genotypic data,
usually an marker data object.}

\item{merge.by}{a character string or name of the column(s) in
\code{phenoData} and \code{namObj} to merge the phenotypic and
genotypic data sets.}

\item{aped}{a data frame for the inverse numerator realtionship
(either pedigree or marker based) of the parents in the NAM population.
The default is NULL.}

\item{prop}{specifies the proportion of the sum of the eigenvalues of the
genomic matrix MM'. Default is prop=1.}

\item{res}{is a logical, with FALSE meaning a "residual" term is
not included, while TRUE includes such a term. The default is FALSE,}

\item{eq.psi}{is a logical.  If it is TRUE a simple scaled identity residual term is
used otherwise a diagonal form is used.  The default FALSE.}

\item{TypeI}{a numerical value determining the level of
significance for detecting a QTL. The default is 0.05.}

\item{attempts}{An integer representing the number of attempts at
convergence for the QTL model. The default is 5.}

\item{data.name}{character string that represents the name of the
data frame for the final model fit using mpwgaim.  If no data.name
is specified, the file is saved in the working directory as the
name of the response dot data.}

\item{trace}{An automatic tracing facility. If \code{trace = TRUE}
then all \code{asreml} output is piped to the screen during the
analysis. If \code{trace = "file.txt"}, then output from all asreml
models is piped to "file.txt". Both trace mechanisms will
display a message if a QTL is detected.}

\item{verboseLev}{numerical value, either 0 or 1, determining the
level of tracing outputted during execution of the algorithm.  A 0
value will produce the standard model fitting output from the
fitted ASReml models involved in the forward selection. A value of
1 will add a table of interval outlier statistics for each
iteration.}

\item{\ldots}{Any other extra arguments to be passed to each of
the \code{asreml} calls. These may also include \code{asreml.control}
arguments.}
}
\value{
An object of class \code{wgnam} which also inherits the
class \code{mpwgaim} and \code{asreml} by default. The object returned is actually an
\code{asreml} object with the addition of
components from the QTL detection.
}
\description{
A computationally efficient whole genome approach to detecting and estimating significant QTL
in Nested Association Mapping Populations using the flexible linear mixed modelling functionality of ASReml-R.
It includes code to handle
high-dimensional problems.  The size of computations in model
fitting is the number of genetic lines rather than the number of
markers * the number of founders. The use of wgnam requires a base model to
be fitted which includes all genetic and non-genetic terms that
are relevant for an analysis without marker information. The
marker information must be contained in an nam object
(\code{namObj} and these effects are added to the phenotypic data
(\code{phenoData}) to allow fitting of models with QTL. Note that
analysis using markers positions is used. All associations
are random effects in the final model.
}
\details{
Method wgnam.asreml
}
\examples{
\dontrun{
wgnam_qtl <- wgnam.asreml(aseModel, phenoData, namObj, merge.by = "Genotype", TypeI = 0.05)
}
}
\references{
Paccapelo, M. V., Kelly, A. M., Christopher, J. T. and
Verbyla, A. P. (2021).  WGNAM: Whole genome nested association mapping.
(under review) Theoretical and Applied Genetics.

Verbyla, A. P., George, A. W., Cavanagh, C. C. and
Verbyla, K. L. (2014).  Whole genome QTL analysis for MAGIC.
Theoretical and Applied Genetics.  DOI:10.1007/s00122-014-2337-4.
}
\seealso{
\code{\link{summary.wgnam}}
}
\author{
Ari Verbyla
}
